:recycle: Move websocket mechanics to separate file
